fix minor issues in cpp02

+++ cpp02/ex01/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: bnaji <bnaji@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/10 19:00:26 by bnaji             #+#    #+#             */
-/*   Updated: 2022/05/12 13:09:19 by bnaji            ###   ########.fr       */
+/*   Updated: 2022/05/17 22:06:02 by bnaji            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ Fixed::Fixed( int const num ) {
 
 Fixed::Fixed( float const num ) {
 	std::cout << "Float constructor called\n";
-	this->_num = round(num * (1 << this->_fb));
+	this->_num = roundf(num * (1 << this->_fb));
 }
 
 Fixed::Fixed( const Fixed & src )
@@ -76,7 +76,7 @@ std::ostream &			operator<<( std::ostream & o, Fixed const & i )
 
 
 float	Fixed::toFloat( void ) const {
-	return (this->_num / (1 << this->_fb));
+	return ((float)this->_num / (float)(1 << this->_fb));
 }
 
 int		Fixed::toInt( void ) const {
